+++ .phan/plugins/InvokePHPNativeSyntaxCheckPlugin.php
@@ -221,7 +221,11 @@ class InvokeExecutionPromise
             //
             // e.g. `""C:\php 7.4.3\php.exe" --syntax-check --no-php-ini < "C:\some project\test.php""`
             // gets unescaped as `"C:\php 7.4.3\php.exe" --syntax-check --no-php-ini < "C:\some project\test.php"`
-            $process = proc_open("\"$cmd\"", $descriptorspec, $pipes);
+            if (PHP_VERSION_ID < 80000) {
+                // In PHP 8.0.0, proc_open started always escaping arguments with additional quotes, so doing it twice would be a bug.
+                $cmd = "\"$cmd\"";
+            }
+            $process = proc_open("$cmd", $descriptorspec, $pipes);
             if (!is_resource($process)) {
                 $this->done = true;
                 $this->error = "Failed to run proc_open in " . __METHOD__;
@@ -229,12 +233,16 @@ class InvokeExecutionPromise
             }
             $this->process = $process;
         } else {
-            $cmd = escapeshellarg($binary) . ' --syntax-check --no-php-ini';
+            $cmd = [$binary, '--syntax-check', '--no-php-ini'];
+            if (PHP_VERSION_ID < 70400) {
+                $cmd = implode(' ', array_map('escapeshellarg', $cmd));
+            }
             $descriptorspec = [
                 ['pipe', 'rb'],
                 ['pipe', 'wb'],
             ];
             $this->binary = $binary;
+            // @phan-suppress-next-line PhanPartialTypeMismatchArgumentInternal
             $process = proc_open($cmd, $descriptorspec, $pipes);
             if (!is_resource($process)) {
                 $this->done = true;