+++ src/Phan/CLI.php
@@ -72,7 +72,7 @@ class CLI
     /**
      * This should be updated to x.y.z-dev after every release, and x.y.z before a release.
      */
-    public const PHAN_VERSION = '2.7.0';
+    public const PHAN_VERSION = '2.7.1-dev';
 
     /**
      * List of short flags passed to getopt